fix(farm): cast growth stage days for stable scanning

+++ backend/internal/farm/catalog_service.go
@@ -88,7 +88,7 @@ func (s *CatalogService) ListVarieties(ctx context.Context, cropID string) ([]Cr
 }
 
 func (s *CatalogService) ListGrowthStages(ctx context.Context, cropID string) ([]CropGrowthStage, error) {
-	rows, err := s.db.Query(ctx, `SELECT id, crop_id, name, sequence_no, min_days, max_days, COALESCE(description, '') FROM crop_growth_stages WHERE crop_id = $1 ORDER BY sequence_no`, cropID)
+	rows, err := s.db.Query(ctx, `SELECT id, crop_id, name, sequence_no, min_days::bigint, max_days::bigint, COALESCE(description, '') FROM crop_growth_stages WHERE crop_id = $1 ORDER BY sequence_no`, cropID)
 	if err != nil {
 		return nil, err
 	}
@@ -96,7 +96,7 @@ func (s *CatalogService) ListGrowthStages(ctx context.Context, cropID string) ([
 	stages := make([]CropGrowthStage, 0)
 	for rows.Next() {
 		var stage CropGrowthStage
-		var minDays, maxDays *int32
+		var minDays, maxDays *int64
 		if err := rows.Scan(&stage.ID, &stage.CropID, &stage.Name, &stage.SequenceNo, &minDays, &maxDays, &stage.Description); err != nil {
 			return nil, err
 		}
